Move common code to AbstractValidationIT
